backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.2.9
django-allauth==65.0.2
django-cors-headers==4.3.1
django-debug-toolbar==4.4.6
django-filter==24.3
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.6.0
